refactor: post, reply 패키징 분리
