+++ src/main/java/com/daily/themoti/community/exception/NoSuchPostExist.java
@@ -1,4 +1,4 @@
-package com.daily.themoti.global.exception;
+package com.daily.themoti.community.exception;
 
 public class NoSuchPostExist extends RuntimeException {
