+++ src/main/java/com/daily/themoti/community/post/dto/PostUpdateRequestDto.java
@@ -1,4 +1,4 @@
-package com.daily.themoti.community.dto;
+package com.daily.themoti.community.post.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
