+++ src/main/java/com/daily/themoti/community/post/repository/PostRepository.java
@@ -1,6 +1,6 @@
-package com.daily.themoti.community.repository;
+package com.daily.themoti.community.post.repository;
 
-import com.daily.themoti.community.domain.Post;
+import com.daily.themoti.community.post.domain.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
